requirements.txt
Flask==2.0.2
httpx==0.20.0
mvc-flask==2.0.0
pytest==6.2.5
python-dotenv==0.19.1
